Allow user to specify title and postid within the blog post.
Currently only works with RST. Markdown is broken.
Add the following to the top of your blog file;

:Title: my own title
:PostID: 123

Title is mandatory. PostID is optional.
Currently it does not check the format of the title nor postid. The title is
expected to be simple text, no markup. PostID must be a single word.